requirements.txt
# blinker==1.9.0
# click==8.1.7
# colorama==0.4.6
# Flask==3.1.0
# Flask-SQLAlchemy==3.1.1
# gunicorn==23.0.0
# itsdangerous==2.2.0
# Jinja2==3.1.4
# MarkupSafe==3.0.2
# packaging==24.2
# SQLAlchemy==2.0.36
# typing_extensions==4.12.2
# Werkzeug==3.1.3
# # Added dependencies
# Flask-JWT-Extended==4.7.0
# Flask-Swagger-UI==4.11.1
# python-dotenv==1.0.0
# psycopg2-binary==2.9.7
# alembic==1.14.0
# flask-migrate==4.0.7
# Flask-Cors==5.0.0

blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
Flask-JWT-Extended==4.7.0
Flask-Swagger-UI==4.11.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
alembic==1.14.0
flask-migrate==4.0.7
Flask-Cors==5.0.0
greenlet==3.1.1
importlib_metadata==8.5.0
Mako==1.3.7
PyJWT==2.10.1
zipp==3.21.0
